chat-bot/requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
efficientnet==1.1.0
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0rc1
multidict==4.7.6
mysql-connector-python==8.0.21
networkx==2.4
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.0a2
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
six==1.15.0
telepot==12.7
telegram==0.0.1
termcolor==1.1.0
tifffile==2020.7.4
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0